refactor: rename files and move ConfigureCmds to Logger
