+++ src/commander/DiscoverDevicesCmd.ts
@@ -1,4 +1,4 @@
-import { ConfigureCmds } from './utils';
+import { ConfigureCmds } from '../shared/Logger';
 import { GetBindingFromContainer } from '../infra/container';
 import { unlinkSync } from 'fs';
 import DiscoverDevicesCase from '../modules/Discovery/DiscoverDevices/DiscoverDevicesCase';
