+++ src/commander/ToggleCmd.ts
@@ -1,5 +1,5 @@
 import { CommandList } from '../infra/enums';
-import { ConfigureCmds } from './utils';
+import { ConfigureCmds } from '../shared/Logger';
 import { GetBindingFromContainer } from '../infra/container';
 import DiscoverDevicesCase from '../modules/Discovery/DiscoverDevices/DiscoverDevicesCase';
 import ReceiveCommandCase from '../modules/Yeelight/ReceiveCommand/ReceiveCommandCase';
