--- src/commander/utils.ts
@@ -1,6 +1,0 @@
-import { logger } from '../shared/Logger';
-
-export const ConfigureCmds = (logLevel?: 'verbose' | 'debug' | 'info') => {
-  logger.level = logLevel;
-};
-
