+++ src/infra/yeelight/devices/ColorFlowExpression.ts
